midterm/requirements.txt
aws-cdk-lib==2.165.0
constructs>=10.0.0,<11.0.0
boto3>=1.20.0
aws-cdk.core==1.134.0  # or another version in the 1.x series
aws-cdk.aws-s3==1.134.0
aws-cdk.aws-lambda==1.134.0
aws-cdk.aws-dynamodb==1.134.0
aws-cdk.aws-events==1.134.0
aws-cdk.aws-events-targets==1.134.0
aws-cdk.aws-s3-notifications==1.134.0